Updated Tutorial, Changed Manager Index Page and More...
Updated Tutorial with more test events to be more reflective of the
actual software.  Changed the color of the 'today' cell to blue.
Implemented finalizeMonth which allows the admin to finalize the
schedule for a given month and make it available to employees. Allowed
for dynamic setting of the slotMinute option in the calendar, and
created the renderCalendar function. Moved js from adminMenu to the
adminCalendar. Fixed bugs in availabilityCalendar. Moved the image
outside the application.